check connection imroved - npe on blockedChannel video list fixed
