+++ app/src/main/java/free/rm/skytube/businessobjects/YouTube/Tasks/GetYouTubeVideosTask.java
@@ -75,7 +75,7 @@ public class GetYouTubeVideosTask extends AsyncTaskParallel<Void, Void, List<You
 
     @Override
     protected List<YouTubeVideo> doInBackground(Void... params) {
-        List<YouTubeVideo> videosList = null;
+        List<YouTubeVideo> videosList;
         ArrayList<YouTubeVideo> youTubeVideoList = new ArrayList<>();
         List<String> blockedChannelIds = new ArrayList<>();
         final BlockedChannelsDb blockedChannelsDb = BlockedChannelsDb.getBlockedChannelsDb();
@@ -98,9 +98,12 @@ public class GetYouTubeVideosTask extends AsyncTaskParallel<Void, Void, List<You
                 }
                 //filtering system to get the videos that are not blocked
                 //videos are checked by their IDs - if their IDs are blocked they are not loaded.
-                for (YouTubeVideo video : videosList) {
-                    if (!blockedChannelIds.contains(video.getChannelId())) {
-                        youTubeVideoList.add(video);
+                // videoList needs to be null checked - if there is not connection we get null point exception
+                if (videosList != null){
+                    for (YouTubeVideo video : videosList) {
+                        if (!blockedChannelIds.contains(video.getChannelId())) {
+                            youTubeVideoList.add(video);
+                        }
                     }
                 }
             } catch (Exception e) {
