+++ app/src/main/java/free/rm/skytube/gui/activities/MainActivity.java
@@ -23,7 +23,10 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.database.Cursor;
+import android.os.AsyncTask;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AlertDialog;
@@ -35,6 +38,7 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.FrameLayout;
+import android.widget.Toast;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -86,6 +90,24 @@ public class MainActivity extends AppCompatActivity implements MainActivityListe
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
+
+		//Asynctask to check network connection on background thread with a handler to reach to UI for the toast.
+		AsyncTask.execute(new Runnable() {
+			@Override
+			public void run() {
+				if (!SkyTubeApp.isInternetAvailable()){
+					Looper.prepare();
+					Handler handler = new Handler();
+					handler.post(new Runnable() {
+						@Override
+						public void run() {
+							Toast.makeText(MainActivity.this, "No internet connection", Toast.LENGTH_LONG).show();
+						}
+					});
+					Looper.loop();
+				}
+			}
+		});
 
 		// check for updates (one time only)
 		if (!updatesCheckerTaskRan) {
