+++ app/src/main/java/free/rm/skytube/gui/fragments/YouTubePlayerFragment.java
@@ -8,8 +8,10 @@ import android.graphics.Rect;
 import android.media.AudioManager;
 import android.media.MediaPlayer;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Handler;
+import android.os.Looper;
 import android.preference.PreferenceManager;
 import android.support.v7.app.AlertDialog;
 import android.util.Log;
@@ -129,6 +131,7 @@ public class YouTubePlayerFragment extends ImmersiveModeFragment implements Medi
 
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+
 		// if immersive mode is enabled then hide the navigation bar
 		if (userWantsImmersiveMode()) {
 			hideNavigationBar();
